apps/scan-engine/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
requests
PyYAML
prance
openapi-spec-validator

